requirements/base.txt
django==1.11

# Configuration
django-environ==0.4.4

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.35

# PostgreSQL adapter for the Python
psycopg2==2.7

# translations
django-rosetta==0.8

# Images
Pillow==5.0

# Django-Rules
rules==1.3

# REST API
# djangorestframework==3.5
djangorestframework==3.7
# djangorestframework-gis==0.10.1
djangorestframework-gis==0.12

# django-rest-serializer-field-permissions==1.0.0
git+https://github.com/hacklabr/django-rest-serializer-field-permissions#egg=django-rest-serializer-field-permissions

rest-framework-cache==0.1
python-memcached

rest-pandas==0.5.0
xlwt==1.1.2

django-compressor==2.1
libsass==0.9.3
django-libsass==0.6

# django-moderation==0.5.0
git+https://github.com/hacklabr/django-moderation#egg=django-moderation

# Django Rules (rules for permissions)
rules==1.3

# Bidirectional UTM-WGS84 coordinates converter for python, used only in archaeological layer import
utm==0.4

# django-spillway deps
# django spillway for protobuf tile layer
# django-spillway==0.7.0
git+https://github.com/hacklabr/django-spillway.git

# GDAL: Required by Fiona and spillway
GDAL==2.1

mapbox-vector-tile==1.2.0

# end of django-spillway

# shapefile export
Fiona==1.7.11

django-map-widgets==0.1.7

# Redis support
django-redis==4.8.0
redis>=2.10.5

# Flatpages admin WYSIWYG
django-ckeditor==5.3.1

# Protected areas
git+https://github.com/hacklabr/geodjango-protected-areas#egg=geodjango-protected-areas

# Administrative boundaries
git+https://github.com/hacklabr/geodjango-boundaries#egg=geodjango-boundaries

# Hydrography
# -e git+https://github.com/hacklabr/geodjango-hydrography#egg=geodjango-hydrography
